# Cython Performance Optimizations