--- a/cvbsdecode/process.py
+++ b/cvbsdecode/process.py
@@ -66,16 +66,9 @@ class FieldCVBSShared:
             )
             # Make sore to not move backwards here
             return None, None, max(line0loc - (meanlinelen * 20), self.inlinelen)
-        
-        validpulses_arr = np.sort(np.asarray(
-            [
-                p[1].start
-                for p in validpulses
-            ], dtype=np.int32
-        ))
 
         linelocs, lineloc_errs, last_validpulse = sync.valid_pulses_to_linelocs(
-            validpulses_arr,
+            validpulses,
             line0loc,
             0,
             meanlinelen,