--- a/vhsdecode/field.py
+++ b/vhsdecode/field.py
@@ -685,16 +685,9 @@ class FieldShared:
                 )
                 ldd.logger.info("lastline < proclines , skipping a tiny bit")
             return None, None, max(line0loc - (meanlinelen * 20), self.inlinelen)
-        
-        validpulses_arr = np.sort(np.asarray(
-            [
-                p[1].start
-                for p in validpulses
-            ], dtype=np.int32
-        ))
 
         linelocs, lineloc_errs, last_validpulse = sync.valid_pulses_to_linelocs(
-            validpulses_arr,
+            validpulses,
             first_hsync_loc,
             first_hsync_loc_line,
             meanlinelen,
@@ -703,6 +696,7 @@ class FieldShared:
             1.9
         )
 
+        lineloc_errs = np.full(proclines, 0, dtype=np.int32)
         self.linelocs0 = linelocs.copy()
 
         # ldd.logger.info("line0loc %s %s", int(line0loc), int(self.meanlinelen))