## Fixes
* Remove `-1` assignment to initial hsync location. This should be set to 0 when not yet filled in. This will fix a possible scenario where the estimated hsync drifts over time if there is a long span of time with no vblanking intervals.

## Performance
Various performance tweaks to the cython sync code. This should help a bit with performance, especially with computers with CPUs that have modern vector instructions.

* Added C compiler flags to optimize more, i.e. `-O3 -march=native -flto`
* Refactor to use `nogil` where possible
* Use contiguous arrays to allow for better auto-vectorization
* Update types to reduce casting

Before (current vhs_decode branch):
```
$ sudo perf stat -e instructions:u,fp_arith_inst_retired.128b_packed_single:u,fp_arith_inst_retired.128b_packed_double:u,fp_arith_inst_retired.256b_packed_single:u,fp_arith_inst_retired.256b_packed_double:u,fp_arith_inst_retired.512b_packed_single:u,fp_arith_inst_retired.512b_packed_double:u -- ./vhs-decode -l 50 -n -t 3 -f 40 --overwrite [file_location].flac test
File Frame 50: VHS                                                              
Completed: saving JSON and exiting.  Took 9.12 seconds to decode 50 frames (5.79 FPS post-setup)

 Performance counter stats for './vhs-decode -l 50 -n -t 3 -f 40 --overwrite [file_location].flac test':

   195,119,171,841      instructions:u                                                        
                 0      fp_arith_inst_retired.128b_packed_single:u                                      
       526,252,196      fp_arith_inst_retired.128b_packed_double:u                                      
             1,783      fp_arith_inst_retired.256b_packed_single:u                                      
        53,500,004      fp_arith_inst_retired.256b_packed_double:u                                      
         4,688,137      fp_arith_inst_retired.512b_packed_single:u                                      
       300,630,107      fp_arith_inst_retired.512b_packed_double:u                                      

      10.112633828 seconds time elapsed

      23.986897000 seconds user
       3.288829000 seconds sys
```

After (this branch):
```
$ sudo perf stat -e instructions:u,fp_arith_inst_retired.128b_packed_single:u,fp_arith_inst_retired.128b_packed_double:u,fp_arith_inst_retired.256b_packed_single:u,fp_arith_inst_retired.256b_packed_double:u,fp_arith_inst_retired.512b_packed_single:u,fp_arith_inst_retired.512b_packed_double:u -- ./vhs-decode -l 50 -n -t 3 -f 40 --overwrite [file_location].flac test
File Frame 50: VHS                                                              
Completed: saving JSON and exiting.  Took 8.83 seconds to decode 50 frames (5.98 FPS post-setup)

 Performance counter stats for './vhs-decode -l 50 -n -t 3 -f 40 --overwrite [file_location].flac test':

   191,981,798,294      instructions:u                                                        
                 0      fp_arith_inst_retired.128b_packed_single:u                                      
       526,252,196      fp_arith_inst_retired.128b_packed_double:u                                      
             1,927      fp_arith_inst_retired.256b_packed_single:u                                      
        53,502,084      fp_arith_inst_retired.256b_packed_double:u                                      
         4,687,993      fp_arith_inst_retired.512b_packed_single:u                                      
       300,630,280      fp_arith_inst_retired.512b_packed_double:u                                      

       9.827384407 seconds time elapsed

      23.535043000 seconds user
       3.147495000 seconds sys
```