@oyvindln 2025-01-03
```CMAKE_BUILD_TYPE``` ```"Release``` means it's a release optimized build as opposed to debug build - which is what you want to use outside of development so it's the wrong thing to test for for adding ```--march=native```. I think it's better to just have a suggestion to use ```CXXFLAGS="-march=native" CFLAGS="-march=native cmake ...." when building locally in the docs instead of putting it into cmakelists.txt. 

LTO already gets enabled earlier in the cmakelists file if supported by the compiler, (```set(CMAKE_INTERPROCEDURAL_OPTIMIZATION_RELEASE ON)```) so don't need to add ```-flto``` in there, we could add -O3 I suppose but not sure if it makes any difference to the default O2 in release mode.


As for the cython side, I think it will break when using MSVC as it is now, not sure if it will pick up CXXFLAGS/CFLAGS from the environment similarly to cmake or if there is some conditional way to add flags only when using gcc/clang, but maybe it's better to just leave it off for now.