@eshaz 2025-01-03
> `CMAKE_BUILD_TYPE` `"Release` means it's a release optimized build as opposed to debug build - which is what you want to use outside of development so it's the wrong thing to test for for adding `--march=native`. I think it's better to just have a suggestion to use ```CXXFLAGS="-march=native" CFLAGS="-march=native cmake ...." when building locally in the docs instead of putting it into cmakelists.txt.
> 
> LTO already gets enabled earlier in the cmakelists file if supported by the compiler, (`set(CMAKE_INTERPROCEDURAL_OPTIMIZATION_RELEASE ON)`) so don't need to add `-flto` in there, we could add -O3 I suppose but not sure if it makes any difference to the default O2 in release mode.

That all sounds good, I'll make these changes.

> As for the cython side, I think it will break when using MSVC as it is now though would have to test to be sure. Not sure if it will pick up CXXFLAGS/CFLAGS from the environment similarly to cmake or if there is some conditional way to add flags only when using gcc/clang, but maybe it's better to just leave it off for now.

Quickly searching this, it looks like there is a way to detect the compiler. I'll update this, and probably just omit the extra flags for MSVC. Someone who can test it / uses Windows can verify if any flags help or harm performance.