@eshaz 2025-01-05
@oyvindln I made the updates discussed above. Let me know if there's anything else that needs changing.